refactor(core): improve code structure
