+++ Cargo.toml
@@ -22,6 +22,7 @@ ratatui = "0.29"
 crossterm = "0.29"
 chrono = "0.4.41"
 ctrlc = { version = "3.4", features = ["termination"] }
+dirs = "6.0.0"
 
 [profile.release]
 opt-level = "z"
